# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Nyx LogBook System API
  version: 1.0.0
  description: Enriched OpenAPI 3.1 specification generated from actual route implementations.
servers:
- url: /
paths:
...
  /admin/learner-learning-paths:
    get:
      summary: GET /admin/learner-learning-paths
      description: Endpoint implemented in src\app\api\admin\learner-learning-paths\route.ts.
      tags:
      - Admin
      responses:
        '401':
          description: Error response
          content:
            application/json:
              schema: &id039
                type: object
                additionalProperties: true
                description: 'TODO: infer response object schema'
        '404':
          description: Error response
          content:
            application/json:
              schema: &id040
                type: object
                additionalProperties: true
                description: 'TODO: infer response object schema'
      security:
      - nextAuthSession: []
      x-roles: &id041
      - ADMIN
    post:
      summary: POST /admin/learner-learning-paths
      description: Endpoint implemented in src\app\api\admin\learner-learning-paths\route.ts.
      tags:
      - Admin
      responses:
        '401':
          description: Error response
          content:
            application/json:
              schema: *id039
        '404':
          description: Error response
          content:
            application/json:
              schema: *id040
      security:
      - nextAuthSession: []
      x-roles: *id041
  /admin/learner-mentor-allocations/{id}:
    put:
      summary: PUT /admin/learner-mentor-allocations/{id}
      description: Endpoint implemented in src\app\api\admin\learner-mentor-allocations\[id]\route.ts.
      tags:
      - Admin
      responses:
        '401':
          description: Error response
          content:
            application/json:
              schema: &id042
                type: object
                additionalProperties: true
                description: 'TODO: infer response object schema'
        '409':
          description: Error response
          content:
            application/json:
              schema: &id043
                type: object
                additionalProperties: true
                description: 'TODO: infer response object schema'
      security:
      - nextAuthSession: []
      x-roles: &id044
      - ADMIN
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: Path parameter id.
    delete:
      summary: DELETE /admin/learner-mentor-allocations/{id}
      description: Endpoint implemented in src\app\api\admin\learner-mentor-allocations\[id]\route.ts.
      tags:
      - Admin
      responses:
        '401':
          description: Error response
          content:
            application/json:
              schema: *id042
        '409':
          description: Error response
          content:
            application/json:
              schema: *id043
      security:
      - nextAuthSession: []
      x-roles: *id044
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: Path parameter id.
  /admin/learner-mentor-allocations:
    get:
      summary: GET /admin/learner-mentor-allocations
      description: Endpoint implemented in src\app\api\admin\learner-mentor-allocations\route.ts.
      tags:
      - Admin
      responses:
        '401':
          description: Error response
          content:
            application/json:
              schema: &id045
                type: object
                additionalProperties: true
                description: 'TODO: infer response object schema'
        '404':
          description: Error response
          content:
            application/json:
              schema: &id046
                type: object
                additionalProperties: true
                description: 'TODO: infer response object schema'
      security:
      - nextAuthSession: []
      x-roles: &id047
      - ADMIN
    post:
      summary: POST /admin/learner-mentor-allocations
      description: Endpoint implemented in src\app\api\admin\learner-mentor-allocations\route.ts.
      tags:
      - Admin
      responses:
        '401':
          description: Error response
          content:
            application/json:
              schema: *id045
        '404':
          description: Error response
          content:
            application/json:
              schema: *id046
      security:
      - nextAuthSession: []
      x-roles: *id047
...
components:
  securitySchemes:
    nextAuthSession:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session cookie used for authenticated requests.